// src/model/DrugModel.tsx
export interface DrugModel {
  _id: number;
  categoryId: number;
  name: string;
  usage: string;
  mechanism: string;
  pharmaco: string;
  noUsage: string;
  warning: string;
  avarez: string;
  tadakhol: string;
  notes: string;
  meghdarMasraf: string;
  ashkal: string;
  isFavorit: boolean;
}